+++ example/lib/main.dart
@@ -6,15 +6,12 @@ import 'package:bloc/bloc.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:hydrated_bloc/hydrated_bloc.dart';
 
-import 'plugins/desktop/desktop.dart';
-
 void main() async {
   // https://github.com/flutter/flutter/pull/38464
   // Changes in Flutter v1.9.4 require you to call WidgetsFlutterBinding.ensureInitialized()
   // before using any plugins if the code is executed before runApp.
   // As a result, you will need the following line if you're using Flutter >=1.9.4.
   WidgetsFlutterBinding.ensureInitialized();
-  setTargetPlatformForDesktop();
   BlocSupervisor.delegate = await HydratedBlocDelegate.build();
   runApp(App());
 }